minimal fix for Qrule/default filtering being applied even when a VQSR model is specified - STARKA-149
Missing "else" restored in calibration_models::clasify_site(indel_info& ii).  This is the primary bug.  It may be desirable to also clean up some special casing, but that should probably be handled as another issue.  Testing should be done on a variety of sample types as the presence of the bug means that previous tests (since Aug 3, 2014) of VQSR performance were not representative of what we will see with the change.