add ref indel error multiplier to reduce undercalling of homozygous indels -- STARKA 152
Add a command-line controllable multiplier for the reference (indel) error probability
(the probability of generating the reference sequence in a read derived from a non-reference allele).
Default value is 1, for backwards compatibility, but evaluation indicates that a value of 100
substantially decreases undercalling without substantially increasing overcalling, at least for
unfiltered calls.

Background: Homozygous indels, especially but not exclusively homopolymer length changing indels,
are 'undercalled' as heterozygous indels at a substantial rate. The reverse, 'overcalling' a truly
heterozygous indel as a homozygous indel, is much more rare (order of 20000 vs 200 events in the
whole genome). While there is no principled reason for the reference indel error probability to
be greater than the equivalent non-reference indel error, there is a detection bias for different
genotypes as follows: Heterozygous indels are relatively easy to detect: typically, several reads
support each allele, so a few reads with errors will not prevent the correct call. In contrast, a
few reference errors are enough to make a true homozygous indel look like a heterozygous locus.
Increasing the reference error rate makes a homozygous indel call a bit easier to make.

Caveat: In preliminary tests, increasing the ref error probability significantly decreased
undercalling in unfiltered calls and under Qrule, but had a markedly smaller effect on
VQSR-filtered calls. This may need investigation. Or, we may need to use a two-level/two-stage
VQSR filtering, for which preliminary results look promising; however, this approach is not
favored as a) it only affects filtering status (i.e. it does not correct any errors we currently
make) and b) it seems like it might not be robust to differences among platforms or chemistries.