STREL-27 refactor indel candidacy test
refactor indel candidate test to decouple the two step procedure:
(1) get expected error rate of this indel
(2) determine significance of observations wrt error rate from (1)

Prior to refactoring the same signifiance test logic was being
repeated, leading to (slight) divergence of the logic.

Note this is not a strict refactor: for indels in homopolymers we
now apply a minimum observation floor of two counts before the indel
can be tested for significance. Previously the floor was being
applied to non-homopolymers only.