STREL-36 fix breakend check for indel scoring
Fix test to determine if a read overlaps an indel breakpoint
by enough bases to be counted towards the indel's score. The
fixed test is now applied uniformly to reads which
(1) support the scored indel
(2) support the reference
(3) support an indel orthogonal to the scored indel

A second test of breakpoint overlap support later in the indel
scoring function has been removed. Now that the first
test is correctly applied to all cases, the second test is no
longer needed (also, the second test was not always valid, but
that doesn't matter now that we don't need it).