STREL-47 restruct error analyzer for new var types
This edit includes some of the larger class shuffling
required to run SNV error analysis in STREL-47. The
inital set of file/class rename changes have been pushed
down to this edit ahead of the rest of the STREL-47
branch to reduce merge conflicts with continuing indel
analysis improvement efforts.