STREL-158 fix rare indel ordering issue
Very rare circumstances cause an allele to be filtered out of the
top allele set but recovered (out of order) as a forced gt allele.

This patch sorts the indel_buffer in indel overlapper in case of
such an event. We should also look into preventing this call in
the first place (better treatment of forced call "other" list).