STREL-227 shift away from legacy naming schemes
This edit removes research content from the installation bin
(most RD workflows are moved to libexec now), and implements the
new workflow naming scheme to reduce confusing legacy/code names
(like 'starling').

The renaming is extended throughout the python workflow code to make
it easier for any new user to jump in, but all compiled code is left
as is.