STREL-336 Fix failed somatic indel normalization
A series of improperly normalized indels needed to be shifted, then
merged, then shifted again to achieve full normalization for the
failed example case. This edit changes the indel normalizer to take
an iterative approach so that any number of required cycles is
normalized. The motivating failure case is captured in a new unit
test on this commit.