STREL-140 Add new EVS feature for RNA-Seq
Add new EVS features which measure mean distance from then read edge for
all alt allele observations. One such feature is added for SNVs and one
for indels. This might be useful in a number of applications, but is
intended to help discriminate RNA-Seq false positive calls where the
end of the read extends into the intron by several bases.