STREL-519 Fix callRegions thread utilization
Remove popen tabix calling loop activated by the callRegions option
in the workflow's task specification thread. At very high thread
counts this call seems to be responsible for stalling the thread to
the point where worker threads are underutilized because task
specification can't keep up. By eliminating this check, we now rely
on strelka itself to detect if its genome segment is not covered in
the callRegion BED file.