STREL-625 Update recommended manta usage
Update README and user guide to reflect our updated recommendation
to use Manta indel candidates for somatic calling only. This change
reflects the role of the local assembler which is now used by the
germline caller. This internal assembler has now obviated the need
for an external indel assembly step.