STREL-651 Fix sequence error counting bounds
Sequence error counting is now bound to the input target region(s) given
for analysis, and counts reporting is not run until all regions have
completed. This ensures that error estimation is not typically impacted
by segment border size, except for the expected case of a large indel
crossing the segment boundary.