STREL-649 Improve somatic EVS feature consistency
Make somatic EVS features conveying QSS_NT and QSI_NT to the EVS model
directly convey the assumption that the normal sample is hom ref by
setting the probability to zero if this is not true. This ensures that
the features use in training and prediciton is consistent with its
intent.