STREL-659 Document read realignment
Document read realignment details. Also (1) Introduce a mechanismy
by which the methods doc can provide design discussion and a bit of
source implementation detail, that can optionally be excluded when
the methods are rendered, and (2) Add source commends and symbol
renaming to move realignment routines closer to new source style.