STREL-674 Refactor starling_read and read_segment
The original design of these structs assumed that assembly-based realignments
of reads were being entered from an external tool, something that is no
longer done. Given the simpler usage, many unspecified contracts on how
the class needs to be initialized and used can be simplified such that
the class is properly initialized by the constructor and ready to use
after construction without further modifications. This also allows for
the elimination or further access restriction of various class data
structures.

This is not intended to be a comprehensive refactor of the the objects
in question, but mostly a focus on increasing object validity at the
point of construction.

In addition, inline source docs have been added or cleaned up, and the
genome or genomic alignment terminology has been replaced with input
alignment.