STREL-830 Add new vcf utility script
Add a script to add or refresh the 'NoPassedVariantGTs' filter on
Strelka's germline VCF output. This filter should be applied to
record which does not have a passing variant genotype in any sample.

Such a filter is already applied to Strelka's multi-sample germline
'variants.vcf' output by default, but the script allows the filter
to be refreshed for subsets of samples extracted from the original
variants.vcf output. This script will also remove the filter, as
approriate, when larger sets of samples are merged, but that is not
the primary anticipated use case.